refactor(api): move AccountRecoveryDemand model
