+++ api/src/identity-access-management/domain/models/AccountRecoveryDemand.js
@@ -1,4 +1,16 @@
-class AccountRecoveryDemand {
+export class AccountRecoveryDemand {
+  /**
+   * @param {{
+   *   id: string,
+   *   userId: string,
+   *   organizationLearnerId: string,
+   *   oldEmail: string,
+   *   newEmail: string,
+   *   temporaryKey: string,
+   *   used: boolean,
+   *   createdAt: string|Date,
+   * }} data
+   */
   constructor({ id, userId, organizationLearnerId, oldEmail, newEmail, temporaryKey, used, createdAt } = {}) {
     this.id = id;
     this.organizationLearnerId = organizationLearnerId;
@@ -10,5 +22,3 @@ class AccountRecoveryDemand {
     this.createdAt = createdAt;
   }
 }
-
-export { AccountRecoveryDemand };
