+++ api/src/identity-access-management/infrastructure/repositories/account-recovery-demand.repository.js
@@ -2,8 +2,8 @@ import _ from 'lodash';
 
 import { knex } from '../../../../db/knex-database-connection.js';
 import { NotFoundError } from '../../../../lib/domain/errors.js';
-import { AccountRecoveryDemand } from '../../../../lib/domain/models/AccountRecoveryDemand.js';
 import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
+import { AccountRecoveryDemand } from '../../domain/models/AccountRecoveryDemand.js';
 
 const _toDomain = (accountRecoveryDemandDTO) => {
   return new AccountRecoveryDemand(accountRecoveryDemandDTO);
