+++ api/tests/identity-access-management/integration/infrastructure/repositories/account-recovery-demand.repository.test.js
@@ -1,8 +1,8 @@
 import lodash from 'lodash';
 
 import { NotFoundError } from '../../../../../lib/domain/errors.js';
-import { AccountRecoveryDemand } from '../../../../../lib/domain/models/AccountRecoveryDemand.js';
 import { DomainTransaction } from '../../../../../lib/infrastructure/DomainTransaction.js';
+import { AccountRecoveryDemand } from '../../../../../src/identity-access-management/domain/models/AccountRecoveryDemand.js';
 import * as accountRecoveryDemandRepository from '../../../../../src/identity-access-management/infrastructure/repositories/account-recovery-demand.repository.js';
 import { catchErr, databaseBuilder, domainBuilder, expect, knex } from '../../../../test-helper.js';
 const { omit } = lodash;
